requirements.txt
mne==1.6.1
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
matplotlib==3.8.2
pyyaml==6.0.1
scipy==1.12.0